requirements.txt
beautifulsoup4==4.12.3
config_file==0.13.1
gensim
kafka_python==2.0.2
keras==3.7.0
nltk==3.9.1
numpy
pandas==2.2.3
PyPDF2==3.0.1
pyfiglet==1.0.2
pyspark==3.5.3
python-dotenv==1.0.1
python-telegram-bot==21.7
Requests==2.32.3
rich==13.9.4
Telethon==1.38.1
tensorflow==2.18.0
textblob==0.18.0.post0
tweepy==4.14.0
vaderSentiment==3.3.2
wordcloud==1.9.4
google-generativeai